Add `bootstrap.py` script which takes care of post-checkout tasks
This script is only present is git checkout of emscripten and not in
emsdk or other packaged versions.

When run with no arguments this script ensures that all the
post-checkout steps required by emscripten have been run and are
up-to-date with the current checkout.  It does this by comparing
`.stamp` files in the `out` directory with the timestamps of inputs
files such as `package.json`.

When `emcc` is, and bootstrap.py is present it is run in `check` mode
which will cause the compiler to error if the one or more of the
bootstrap commands needs to be run (e.g. if package.json was modified).

This first thing this allows us to do is delete all the checked in
launcher files from the top level directory and create them on first
use.  This means we can create just the `.bat` file launchers on windows
and just the `.sh` launchers on non-windows.  The seemly small step
moves is towards have a much cleanup top-level directory.

Note that I have left the launchers for emcc checked-in (at least for
the transition period) so that new users will see the error message and
know to run `bootstrap.py` when they first perform that checkout.